+++ listen-interface/index.html
@@ -2,13 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/listen-more.png" />
+    <link rel="icon" type="image/svg+xml" href="/listen-new.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <!-- Force browsers to check for updates -->
     <meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate" />
     <meta http-equiv="Pragma" content="no-cache" />
     <meta http-equiv="Expires" content="0" />
-    <title>listen</title>
+    <title>Listen</title>
   </head>
   <body>
     <div id="root"></div>
